@@ -1,19 +1,19 @@
 name: flutter_quill
 description: A rich text editor supporting mobile and web (Demo App @ bulletjournal.us)
-version: 2.0.23
+version: 3.8.5
 #author: bulletjournal
 homepage: https://bulletjournal.us/home/index.html
 repository: https://github.com/singerdmx/flutter-quill
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  flutter: ">=2.5.3"
 
 dependencies:
   flutter:
     sdk: flutter
   collection: ^1.15.0
-  flutter_colorpicker: ^0.6.0
+  flutter_colorpicker: ^1.0.3
   flutter_keyboard_visibility: ^5.0.0
   image_picker: ^0.8.2
   photo_view: ^0.13.0
@@ -27,6 +27,7 @@ dependencies:
   youtube_player_flutter: ^8.0.0
   diff_match_patch: ^0.4.1
   i18n_extension: ^4.1.3
+  gallery_saver: ^2.3.2
 
 dev_dependencies:
   flutter_test:
