@@ -2,7 +2,7 @@ library flutter_quill;
 
 export 'src/models/documents/attribute.dart';
 export 'src/models/documents/document.dart';
-export 'src/models/documents/nodes/embed.dart';
+export 'src/models/documents/nodes/embeddable.dart';
 export 'src/models/documents/nodes/leaf.dart';
 export 'src/models/quill_delta.dart';
 export 'src/models/themes/quill_dialog_theme.dart';
@@ -10,5 +10,6 @@ export 'src/models/themes/quill_icon_theme.dart';
 export 'src/widgets/controller.dart';
 export 'src/widgets/default_styles.dart';
 export 'src/widgets/editor.dart';
+export 'src/widgets/link.dart' show LinkActionPickerDelegate, LinkMenuAction;
 export 'src/widgets/style_widgets/style_widgets.dart';
 export 'src/widgets/toolbar.dart';
