@@ -2,6 +2,7 @@ import '../documents/attribute.dart';
 import '../quill_delta.dart';
 import 'rule.dart';
 
+/// A heuristic rule for format (retain) operations.
 abstract class FormatRule extends Rule {
   const FormatRule();
 
@@ -16,6 +17,8 @@ abstract class FormatRule extends Rule {
   }
 }
 
+/// Produces Delta with line-level attributes applied strictly to
+/// newline characters.
 class ResolveLineFormatRule extends FormatRule {
   const ResolveLineFormatRule();
 
@@ -26,69 +29,81 @@ class ResolveLineFormatRule extends FormatRule {
       return null;
     }
 
-    var delta = Delta()..retain(index);
+    // Apply line styles to all newline characters within range of this
+    // retain operation.
+    var result = Delta()..retain(index);
     final itr = DeltaIterator(document)..skip(index);
     Operation op;
     for (var cur = 0; cur < len! && itr.hasNext; cur += op.length!) {
       op = itr.next(len - cur);
-      if (op.data is! String || !(op.data as String).contains('\n')) {
-        delta.retain(op.length!);
+      final opText = op.data is String ? op.data as String : '';
+      if (!opText.contains('\n')) {
+        result.retain(op.length!);
         continue;
       }
-      final text = op.data as String;
-      final tmp = Delta();
-      var offset = 0;
 
-      // Enforce Block Format exclusivity by rule
-      final removedBlocks = Attribute.exclusiveBlockKeys.contains(attribute.key)
-          ? op.attributes?.keys
-                  .where((key) =>
-                      Attribute.exclusiveBlockKeys.contains(key) &&
-                      attribute.key != key &&
-                      attribute.value != null)
-                  .map((key) => MapEntry<String, dynamic>(key, null)) ??
-              []
-          : <MapEntry<String, dynamic>>[];
-
-      for (var lineBreak = text.indexOf('\n');
-          lineBreak >= 0;
-          lineBreak = text.indexOf('\n', offset)) {
-        tmp
-          ..retain(lineBreak - offset)
-          ..retain(1, attribute.toJson()..addEntries(removedBlocks));
-        offset = lineBreak + 1;
-      }
-      tmp.retain(text.length - offset);
-      delta = delta.concat(tmp);
+      final delta = _applyAttribute(opText, op, attribute);
+      result = result.concat(delta);
     }
-
+    // And include extra newline after retain
     while (itr.hasNext) {
       op = itr.next();
-      final text = op.data is String ? (op.data as String?)! : '';
-      final lineBreak = text.indexOf('\n');
-      if (lineBreak < 0) {
-        delta.retain(op.length!);
+      final opText = op.data is String ? op.data as String : '';
+      final lf = opText.indexOf('\n');
+      if (lf < 0) {
+        result.retain(op.length!);
         continue;
       }
-      // Enforce Block Format exclusivity by rule
-      final removedBlocks = Attribute.exclusiveBlockKeys.contains(attribute.key)
-          ? op.attributes?.keys
-                  .where((key) =>
-                      Attribute.exclusiveBlockKeys.contains(key) &&
-                      attribute.key != key &&
-                      attribute.value != null)
-                  .map((key) => MapEntry<String, dynamic>(key, null)) ??
-              []
-          : <MapEntry<String, dynamic>>[];
-      delta
-        ..retain(lineBreak)
-        ..retain(1, attribute.toJson()..addEntries(removedBlocks));
+
+      final delta = _applyAttribute(opText, op, attribute, firstOnly: true);
+      result = result.concat(delta);
       break;
     }
-    return delta;
+    return result;
+  }
+
+  Delta _applyAttribute(String text, Operation op, Attribute attribute,
+      {bool firstOnly = false}) {
+    final result = Delta();
+    var offset = 0;
+    var lf = text.indexOf('\n');
+    final removedBlocks = _getRemovedBlocks(attribute, op);
+    while (lf >= 0) {
+      final actualStyle = attribute.toJson()..addEntries(removedBlocks);
+      result
+        ..retain(lf - offset)
+        ..retain(1, actualStyle);
+
+      if (firstOnly) {
+        return result;
+      }
+
+      offset = lf + 1;
+      lf = text.indexOf('\n', offset);
+    }
+    // Retain any remaining characters in text
+    result.retain(text.length - offset);
+    return result;
+  }
+
+  Iterable<MapEntry<String, dynamic>> _getRemovedBlocks(
+      Attribute<dynamic> attribute, Operation op) {
+    // Enforce Block Format exclusivity by rule
+    if (!Attribute.exclusiveBlockKeys.contains(attribute.key)) {
+      return <MapEntry<String, dynamic>>[];
+    }
+
+    return op.attributes?.keys
+            .where((key) =>
+                Attribute.exclusiveBlockKeys.contains(key) &&
+                attribute.key != key &&
+                attribute.value != null)
+            .map((key) => MapEntry<String, dynamic>(key, null)) ??
+        [];
   }
 }
 
+/// Allows updating link format with collapsed selection.
 class FormatLinkAtCaretPositionRule extends FormatRule {
   const FormatLinkAtCaretPositionRule();
 
@@ -121,6 +136,8 @@ class FormatLinkAtCaretPositionRule extends FormatRule {
   }
 }
 
+/// Produces Delta with inline-level attributes applied too all characters
+/// except newlines.
 class ResolveInlineFormatRule extends FormatRule {
   const ResolveInlineFormatRule();
 
