@@ -2,10 +2,18 @@ import 'package:flutter/rendering.dart';
 
 import '../models/documents/nodes/container.dart';
 
+/// A common interface to render boxes which represent a piece of rich text
+/// content.
+///
+/// See also:
+///   * [RenderParagraphProxy] implementation of this interface which wraps
+///     built-in [RenderParagraph]
+///   * [RenderEmbedProxy] implementation of this interface which wraps
+///     an arbitrary render box representing an embeddable object.
 abstract class RenderContentProxyBox implements RenderBox {
-  double getPreferredLineHeight();
+  double get preferredLineHeight;
 
-  Offset getOffsetForCaret(TextPosition position, Rect? caretPrototype);
+  Offset getOffsetForCaret(TextPosition position, Rect caretPrototype);
 
   TextPosition getPositionForOffset(Offset offset);
 
@@ -13,6 +21,13 @@ abstract class RenderContentProxyBox implements RenderBox {
 
   TextRange getWordBoundary(TextPosition position);
 
+  /// Returns a list of rects that bound the given selection.
+  ///
+  /// A given selection might have more than one rect if this text painter
+  /// contains bidirectional text because logically contiguous text might not be
+  /// visually contiguous.
+  ///
+  /// Valid only after [layout]
   List<TextBox> getBoxesForSelection(TextSelection textSelection);
 }
 
@@ -23,7 +38,7 @@ abstract class RenderContentProxyBox implements RenderBox {
 /// [RenderContentProxyBox].
 abstract class RenderEditableBox extends RenderBox {
   /// The document node represented by this render box.
-  Container getContainer();
+  Container get container;
 
   /// Returns preferred line height at specified `position` in text.
   ///
@@ -119,4 +134,8 @@ abstract class RenderEditableBox extends RenderBox {
   /// Returns the [Rect] in local coordinates for the caret at the given text
   /// position.
   Rect getLocalRectForCaret(TextPosition position);
+
+  /// Returns the [Rect] of the caret prototype at the given text
+  /// position. [Rect] starts at origin.
+  Rect getCaretPrototype(TextPosition position);
 }
