@@ -5,12 +5,13 @@ import 'package:tuple/tuple.dart';
 
 import '../models/documents/attribute.dart';
 import '../models/documents/document.dart';
-import '../models/documents/nodes/embed.dart';
+import '../models/documents/nodes/embeddable.dart';
 import '../models/documents/style.dart';
 import '../models/quill_delta.dart';
-import '../utils/diff_delta.dart';
+import '../utils/delta.dart';
 
 typedef ReplaceTextCallback = bool Function(int index, int len, Object? data);
+typedef DeleteCallback = void Function(int cursorPosition, bool forward);
 
 class QuillController extends ChangeNotifier {
   QuillController({
@@ -18,6 +19,8 @@ class QuillController extends ChangeNotifier {
     required TextSelection selection,
     bool keepStyleOnNewLine = false,
     this.onReplaceText,
+    this.onDelete,
+    this.onSelectionCompleted,
   })  : _selection = selection,
         _keepStyleOnNewLine = keepStyleOnNewLine;
 
@@ -39,9 +42,14 @@ class QuillController extends ChangeNotifier {
   TextSelection get selection => _selection;
   TextSelection _selection;
 
-  /// Manual [replaceText] handler
+  /// Custom [replaceText] handler
   /// Return false to ignore the event
   ReplaceTextCallback? onReplaceText;
+
+  /// Custom delete handler
+  DeleteCallback? onDelete;
+
+  void Function()? onSelectionCompleted;
 
   /// Store any styles attribute that got toggled by the tap of a button
   /// and that has not been applied yet.
@@ -74,6 +82,20 @@ class QuillController extends ChangeNotifier {
     return document
         .collectStyle(selection.start, selection.end - selection.start)
         .mergeAll(toggledStyle);
+  }
+
+  /// Returns all styles for each node within selection
+  List<Tuple2<int, Style>> getAllIndividualSelectionStyles() {
+    final styles = document.collectAllIndividualStyles(
+        selection.start, selection.end - selection.start);
+    return styles;
+  }
+
+  /// Returns plain text for each node within selection
+  String getPlainText() {
+    final text =
+        document.getPlainText(selection.start, selection.end - selection.start);
+    return text;
   }
 
   /// Returns all styles for any character within the specified text range.
@@ -116,6 +138,12 @@ class QuillController extends ChangeNotifier {
   bool get hasUndo => document.hasUndo;
 
   bool get hasRedo => document.hasRedo;
+
+  /// clear editor
+  void clear() {
+    replaceText(0, plainTextEditingValue.text.length - 1, '',
+        const TextSelection.collapsed(offset: 0));
+  }
 
   void replaceText(
       int index, int len, Object? data, TextSelection? textSelection,
@@ -186,14 +214,33 @@ class QuillController extends ChangeNotifier {
     ignoreFocusOnTextChange = false;
   }
 
+  /// Called in two cases:
+  /// forward == false && textBefore.isEmpty
+  /// forward == true && textAfter.isEmpty
+  /// Android only
+  /// see https://github.com/singerdmx/flutter-quill/discussions/514
+  void handleDelete(int cursorPosition, bool forward) =>
+      onDelete?.call(cursorPosition, forward);
+
+  void formatTextStyle(int index, int len, Style style) {
+    style.attributes.forEach((key, attr) {
+      formatText(index, len, attr);
+    });
+  }
+
   void formatText(int index, int len, Attribute? attribute) {
     if (len == 0 &&
         attribute!.isInline &&
         attribute.key != Attribute.link.key) {
+      // Add the attribute to our toggledStyle.
+      // It will be used later upon insertion.
       toggledStyle = toggledStyle.put(attribute);
     }
 
     final change = document.format(index, len, attribute);
+    // Transform selection against the composed change and give priority to
+    // the change. This is needed in cases when format operation actually
+    // inserts data into the document (e.g. embeds).
     final adjustedSelection = selection.copyWith(
         baseOffset: change.transformPosition(selection.baseOffset),
         extentOffset: change.transformPosition(selection.extentOffset));
@@ -205,6 +252,22 @@ class QuillController extends ChangeNotifier {
 
   void formatSelection(Attribute? attribute) {
     formatText(selection.start, selection.end - selection.start, attribute);
+  }
+
+  void moveCursorToStart() {
+    updateSelection(
+        const TextSelection.collapsed(offset: 0), ChangeSource.LOCAL);
+  }
+
+  void moveCursorToPosition(int position) {
+    updateSelection(
+        TextSelection.collapsed(offset: position), ChangeSource.LOCAL);
+  }
+
+  void moveCursorToEnd() {
+    updateSelection(
+        TextSelection.collapsed(offset: plainTextEditingValue.text.length),
+        ChangeSource.LOCAL);
   }
 
   void updateSelection(TextSelection textSelection, ChangeSource source) {
