@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
 import 'package:tuple/tuple.dart';
 
-import 'style_widgets/style_widgets.dart';
+import '../../flutter_quill.dart';
+import '../models/documents/style.dart';
+import '../utils/platform.dart';
 
 class QuillStyles extends InheritedWidget {
   const QuillStyles({
@@ -28,20 +29,95 @@ class QuillStyles extends InheritedWidget {
   }
 }
 
+/// Style theme applied to a block of rich text, including single-line
+/// paragraphs.
 class DefaultTextBlockStyle {
   DefaultTextBlockStyle(
       this.style, this.verticalSpacing, this.lineSpacing, this.decoration,
       {this.indent});
 
+  /// Base text style for a text block.
   final TextStyle style;
 
+  /// Vertical spacing around a text block.
   final Tuple2<double, double> verticalSpacing;
 
+  /// Vertical spacing for individual lines within a text block.
+  ///
   final Tuple2<double, double> lineSpacing;
 
   final double? indent;
 
+  /// Decoration of a text block.
+  ///
+  /// Decoration, if present, is painted in the content area, excluding
+  /// any [spacing].
   final BoxDecoration? decoration;
+}
+
+/// Theme data for inline code.
+class InlineCodeStyle {
+  InlineCodeStyle({
+    required this.style,
+    this.header1,
+    this.header2,
+    this.header3,
+    this.backgroundColor,
+    this.radius,
+  });
+
+  /// Base text style for an inline code.
+  final TextStyle style;
+
+  /// Style override for inline code in header level 1.
+  final TextStyle? header1;
+
+  /// Style override for inline code in headings level 2.
+  final TextStyle? header2;
+
+  /// Style override for inline code in headings level 3.
+  final TextStyle? header3;
+
+  /// Background color for inline code.
+  final Color? backgroundColor;
+
+  /// Radius used when paining the background.
+  final Radius? radius;
+
+  /// Returns effective style to use for inline code for the specified
+  /// [lineStyle].
+  TextStyle styleFor(Style lineStyle) {
+    if (lineStyle.containsKey(Attribute.h1.key)) {
+      return header1 ?? style;
+    }
+    if (lineStyle.containsKey(Attribute.h2.key)) {
+      return header2 ?? style;
+    }
+    if (lineStyle.containsKey(Attribute.h3.key)) {
+      return header3 ?? style;
+    }
+    return style;
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) {
+      return true;
+    }
+    if (other is! InlineCodeStyle) {
+      return false;
+    }
+    return other.style == style &&
+        other.header1 == header1 &&
+        other.header2 == header2 &&
+        other.header3 == header3 &&
+        other.backgroundColor == backgroundColor &&
+        other.radius == radius;
+  }
+
+  @override
+  int get hashCode =>
+      Object.hash(style, header1, header2, header3, backgroundColor, radius);
 }
 
 class DefaultListBlockStyle extends DefaultTextBlockStyle {
@@ -91,7 +167,9 @@ class DefaultStyles {
   final TextStyle? small;
   final TextStyle? underline;
   final TextStyle? strikeThrough;
-  final TextStyle? inlineCode;
+
+  /// Theme of inline code.
+  final InlineCodeStyle? inlineCode;
   final TextStyle? sizeSmall; // 'small'
   final TextStyle? sizeLarge; // 'large'
   final TextStyle? sizeHuge; // 'huge'
@@ -114,20 +192,17 @@ class DefaultStyles {
     );
     const baseSpacing = Tuple2<double, double>(6, 0);
     String fontFamily;
-    switch (themeData.platform) {
-      case TargetPlatform.iOS:
-      case TargetPlatform.macOS:
-        fontFamily = 'Menlo';
-        break;
-      case TargetPlatform.android:
-      case TargetPlatform.fuchsia:
-      case TargetPlatform.windows:
-      case TargetPlatform.linux:
-        fontFamily = 'Roboto Mono';
-        break;
-      default:
-        throw UnimplementedError();
+    if (isAppleOS(themeData.platform)) {
+      fontFamily = 'Menlo';
+    } else {
+      fontFamily = 'Roboto Mono';
     }
+
+    final inlineCodeStyle = TextStyle(
+      fontSize: 14,
+      color: themeData.colorScheme.primaryVariant.withOpacity(0.8),
+      fontFamily: fontFamily,
+    );
 
     return DefaultStyles(
         h1: DefaultTextBlockStyle(
@@ -167,10 +242,19 @@ class DefaultStyles {
         small: const TextStyle(fontSize: 12, color: Colors.black45),
         underline: const TextStyle(decoration: TextDecoration.underline),
         strikeThrough: const TextStyle(decoration: TextDecoration.lineThrough),
-        inlineCode: TextStyle(
-          color: Colors.blue.shade900.withOpacity(0.9),
-          fontFamily: fontFamily,
-          fontSize: 13,
+        inlineCode: InlineCodeStyle(
+          backgroundColor: Colors.grey.shade100,
+          radius: const Radius.circular(3),
+          style: inlineCodeStyle,
+          header1: inlineCodeStyle.copyWith(
+            fontSize: 32,
+            fontWeight: FontWeight.w300,
+          ),
+          header2: inlineCodeStyle.copyWith(fontSize: 22),
+          header3: inlineCodeStyle.copyWith(
+            fontSize: 18,
+            fontWeight: FontWeight.w500,
+          ),
         ),
         link: TextStyle(
           color: themeData.colorScheme.secondary,
