@@ -2,7 +2,8 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:youtube_player_flutter/youtube_player_flutter.dart';
-import '../../flutter_quill.dart';
+
+import '../../../flutter_quill.dart';
 
 class YoutubeVideoApp extends StatefulWidget {
   const YoutubeVideoApp(
