@@ -1,129 +1,89 @@
-import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 
-//fixme workaround flutter MacOS issue https://github.com/flutter/flutter/issues/75595
-extension _LogicalKeyboardKeyCaseExt on LogicalKeyboardKey {
-  static const _kUpperToLowerDist = 0x20;
-  static final _kLowerCaseA = LogicalKeyboardKey.keyA.keyId;
-  static final _kLowerCaseZ = LogicalKeyboardKey.keyZ.keyId;
+class QuillPressedKeys extends ChangeNotifier {
+  static QuillPressedKeys of(BuildContext context) {
+    final widget =
+        context.dependOnInheritedWidgetOfExactType<_QuillPressedKeysAccess>();
+    return widget!.pressedKeys;
+  }
 
-  LogicalKeyboardKey toUpperCase() {
-    if (keyId < _kLowerCaseA || keyId > _kLowerCaseZ) return this;
-    return LogicalKeyboardKey(keyId - _kUpperToLowerDist);
+  bool _metaPressed = false;
+  bool _controlPressed = false;
+
+  /// Whether meta key is currently pressed.
+  bool get metaPressed => _metaPressed;
+
+  /// Whether control key is currently pressed.
+  bool get controlPressed => _controlPressed;
+
+  void _updatePressedKeys(Set<LogicalKeyboardKey> pressedKeys) {
+    final meta = pressedKeys.contains(LogicalKeyboardKey.metaLeft) ||
+        pressedKeys.contains(LogicalKeyboardKey.metaRight);
+    final control = pressedKeys.contains(LogicalKeyboardKey.controlLeft) ||
+        pressedKeys.contains(LogicalKeyboardKey.controlRight);
+    if (_metaPressed != meta || _controlPressed != control) {
+      _metaPressed = meta;
+      _controlPressed = control;
+      notifyListeners();
+    }
   }
 }
 
-enum InputShortcut { CUT, COPY, PASTE, SELECT_ALL, UNDO, REDO }
+class QuillKeyboardListener extends StatefulWidget {
+  const QuillKeyboardListener({required this.child, Key? key})
+      : super(key: key);
 
-typedef CursorMoveCallback = void Function(
-    LogicalKeyboardKey key, bool wordModifier, bool lineModifier, bool shift);
-typedef InputShortcutCallback = void Function(InputShortcut? shortcut);
-typedef OnDeleteCallback = void Function(bool forward);
+  final Widget child;
 
-class KeyboardEventHandler {
-  KeyboardEventHandler(this.onCursorMove, this.onShortcut, this.onDelete);
+  @override
+  QuillKeyboardListenerState createState() => QuillKeyboardListenerState();
+}
 
-  final CursorMoveCallback onCursorMove;
-  final InputShortcutCallback onShortcut;
-  final OnDeleteCallback onDelete;
+class QuillKeyboardListenerState extends State<QuillKeyboardListener> {
+  final QuillPressedKeys _pressedKeys = QuillPressedKeys();
 
-  static final Set<LogicalKeyboardKey> _moveKeys = <LogicalKeyboardKey>{
-    LogicalKeyboardKey.arrowRight,
-    LogicalKeyboardKey.arrowLeft,
-    LogicalKeyboardKey.arrowUp,
-    LogicalKeyboardKey.arrowDown,
-  };
+  bool _keyEvent(KeyEvent event) {
+    _pressedKeys
+        ._updatePressedKeys(HardwareKeyboard.instance.logicalKeysPressed);
+    return false;
+  }
 
-  static final Set<LogicalKeyboardKey> _shortcutKeys = <LogicalKeyboardKey>{
-    LogicalKeyboardKey.keyA,
-    LogicalKeyboardKey.keyC,
-    LogicalKeyboardKey.keyV,
-    LogicalKeyboardKey.keyX,
-    LogicalKeyboardKey.keyZ.toUpperCase(),
-    LogicalKeyboardKey.keyZ,
-    LogicalKeyboardKey.delete,
-    LogicalKeyboardKey.backspace,
-  };
+  @override
+  void initState() {
+    super.initState();
+    HardwareKeyboard.instance.addHandler(_keyEvent);
+    _pressedKeys
+        ._updatePressedKeys(HardwareKeyboard.instance.logicalKeysPressed);
+  }
 
-  static final Set<LogicalKeyboardKey> _nonModifierKeys = <LogicalKeyboardKey>{
-    ..._shortcutKeys,
-    ..._moveKeys,
-  };
+  @override
+  void dispose() {
+    HardwareKeyboard.instance.removeHandler(_keyEvent);
+    _pressedKeys.dispose();
+    super.dispose();
+  }
 
-  static final Set<LogicalKeyboardKey> _modifierKeys = <LogicalKeyboardKey>{
-    LogicalKeyboardKey.shift,
-    LogicalKeyboardKey.control,
-    LogicalKeyboardKey.alt,
-  };
+  @override
+  Widget build(BuildContext context) {
+    return _QuillPressedKeysAccess(
+      pressedKeys: _pressedKeys,
+      child: widget.child,
+    );
+  }
+}
 
-  static final Set<LogicalKeyboardKey> _macOsModifierKeys =
-      <LogicalKeyboardKey>{
-    LogicalKeyboardKey.shift,
-    LogicalKeyboardKey.meta,
-    LogicalKeyboardKey.alt,
-  };
+class _QuillPressedKeysAccess extends InheritedWidget {
+  const _QuillPressedKeysAccess({
+    required this.pressedKeys,
+    required Widget child,
+    Key? key,
+  }) : super(key: key, child: child);
 
-  static final Set<LogicalKeyboardKey> _interestingKeys = <LogicalKeyboardKey>{
-    ..._modifierKeys,
-    ..._macOsModifierKeys,
-    ..._nonModifierKeys,
-  };
+  final QuillPressedKeys pressedKeys;
 
-  static final Map<LogicalKeyboardKey, InputShortcut> _keyToShortcut = {
-    LogicalKeyboardKey.keyX: InputShortcut.CUT,
-    LogicalKeyboardKey.keyC: InputShortcut.COPY,
-    LogicalKeyboardKey.keyV: InputShortcut.PASTE,
-    LogicalKeyboardKey.keyA: InputShortcut.SELECT_ALL,
-  };
-
-  KeyEventResult handleRawKeyEvent(RawKeyEvent event) {
-    if (kIsWeb) {
-      // On web platform, we ignore the key because it's already processed.
-      return KeyEventResult.ignored;
-    }
-
-    if (event is! RawKeyDownEvent) {
-      return KeyEventResult.ignored;
-    }
-
-    final keysPressed =
-        LogicalKeyboardKey.collapseSynonyms(RawKeyboard.instance.keysPressed);
-    final key = event.logicalKey;
-    final isMacOS = event.data is RawKeyEventDataMacOs;
-    if (!_nonModifierKeys.contains(key) ||
-        keysPressed
-                .difference(isMacOS ? _macOsModifierKeys : _modifierKeys)
-                .length >
-            1 ||
-        keysPressed.difference(_interestingKeys).isNotEmpty) {
-      return KeyEventResult.ignored;
-    }
-
-    final isShortcutModifierPressed =
-        isMacOS ? event.isMetaPressed : event.isControlPressed;
-
-    if (_moveKeys.contains(key)) {
-      onCursorMove(
-          key,
-          isMacOS ? event.isAltPressed : event.isControlPressed,
-          isMacOS ? event.isMetaPressed : event.isAltPressed,
-          event.isShiftPressed);
-    } else if (isShortcutModifierPressed && (_shortcutKeys.contains(key))) {
-      if (key == LogicalKeyboardKey.keyZ ||
-          key == LogicalKeyboardKey.keyZ.toUpperCase()) {
-        onShortcut(
-            event.isShiftPressed ? InputShortcut.REDO : InputShortcut.UNDO);
-      } else {
-        onShortcut(_keyToShortcut[key]);
-      }
-    } else if (key == LogicalKeyboardKey.delete) {
-      onDelete(true);
-    } else if (key == LogicalKeyboardKey.backspace) {
-      onDelete(false);
-    } else {
-      return KeyEventResult.ignored;
-    }
-    return KeyEventResult.handled;
+  @override
+  bool updateShouldNotify(covariant _QuillPressedKeysAccess oldWidget) {
+    return oldWidget.pressedKeys != pressedKeys;
   }
 }
