@@ -1,14 +1,16 @@
+import 'dart:ui';
+
+import 'package:flutter/animation.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
 
 import '../../models/documents/document.dart';
-import '../../utils/diff_delta.dart';
+import '../../utils/delta.dart';
 import '../editor.dart';
 
 mixin RawEditorStateTextInputClientMixin on EditorState
     implements TextInputClient {
-  final List<TextEditingValue> _sentRemoteValues = [];
   TextInputConnection? _textInputConnection;
   TextEditingValue? _lastKnownRemoteTextEditingValue;
 
@@ -47,7 +49,7 @@ mixin RawEditorStateTextInputClientMixin on EditorState
     }
 
     if (!hasConnection) {
-      _lastKnownRemoteTextEditingValue = getTextEditingValue();
+      _lastKnownRemoteTextEditingValue = textEditingValue;
       _textInputConnection = TextInput.attach(
         this,
         TextInputConfiguration(
@@ -60,8 +62,8 @@ mixin RawEditorStateTextInputClientMixin on EditorState
         ),
       );
 
+      _updateSizeAndTransform();
       _textInputConnection!.setEditingState(_lastKnownRemoteTextEditingValue!);
-      // _sentRemoteValues.add(_lastKnownRemoteTextEditingValue);
     }
 
     _textInputConnection!.show();
@@ -75,7 +77,6 @@ mixin RawEditorStateTextInputClientMixin on EditorState
     _textInputConnection!.close();
     _textInputConnection = null;
     _lastKnownRemoteTextEditingValue = null;
-    _sentRemoteValues.clear();
   }
 
   /// Updates remote value based on current state of [document] and
@@ -88,12 +89,14 @@ mixin RawEditorStateTextInputClientMixin on EditorState
       return;
     }
 
+    final value = textEditingValue;
+
     // Since we don't keep track of the composing range in value provided
     // by the Controller we need to add it here manually before comparing
     // with the last known remote value.
     // It is important to prevent excessive remote updates as it can cause
     // race conditions.
-    final actualValue = getTextEditingValue().copyWith(
+    final actualValue = value.copyWith(
       composing: _lastKnownRemoteTextEditingValue!.composing,
     );
 
@@ -101,20 +104,15 @@ mixin RawEditorStateTextInputClientMixin on EditorState
       return;
     }
 
-    final shouldRemember =
-        getTextEditingValue().text != _lastKnownRemoteTextEditingValue!.text;
     _lastKnownRemoteTextEditingValue = actualValue;
     _textInputConnection!.setEditingState(
       // Set composing to (-1, -1), otherwise an exception will be thrown if
       // the values are different.
       actualValue.copyWith(composing: const TextRange(start: -1, end: -1)),
     );
-    if (shouldRemember) {
-      // Only keep track if text changed (selection changes are not relevant)
-      _sentRemoteValues.add(actualValue);
-    }
   }
 
+  // Start TextInputClient implementation
   @override
   TextEditingValue? get currentTextEditingValue =>
       _lastKnownRemoteTextEditingValue;
@@ -126,22 +124,6 @@ mixin RawEditorStateTextInputClientMixin on EditorState
   @override
   void updateEditingValue(TextEditingValue value) {
     if (!shouldCreateInputConnection) {
-      return;
-    }
-
-    if (_sentRemoteValues.contains(value)) {
-      /// There is a race condition in Flutter text input plugin where sending
-      /// updates to native side too often results in broken behavior.
-      /// TextInputConnection.setEditingValue is an async call to native side.
-      /// For each such call native side _always_ sends an update which triggers
-      /// this method (updateEditingValue) with the same value we've sent it.
-      /// If multiple calls to setEditingValue happen too fast and we only
-      /// track the last sent value then there is no way for us to filter out
-      /// automatic callbacks from native side.
-      /// Therefore we have to keep track of all values we send to the native
-      /// side and when we see this same value appear here we skip it.
-      /// This is fragile but it's probably the only available option.
-      _sentRemoteValues.remove(value);
       return;
     }
 
@@ -186,9 +168,117 @@ mixin RawEditorStateTextInputClientMixin on EditorState
     // no-op
   }
 
+  // The time it takes for the floating cursor to snap to the text aligned
+  // cursor position after the user has finished placing it.
+  static const Duration _floatingCursorResetTime = Duration(milliseconds: 125);
+
+  // The original position of the caret on FloatingCursorDragState.start.
+  Rect? _startCaretRect;
+
+  // The most recent text position as determined by the location of the floating
+  // cursor.
+  TextPosition? _lastTextPosition;
+
+  // The offset of the floating cursor as determined from the start call.
+  Offset? _pointOffsetOrigin;
+
+  // The most recent position of the floating cursor.
+  Offset? _lastBoundedOffset;
+
+  // Because the center of the cursor is preferredLineHeight / 2 below the touch
+  // origin, but the touch origin is used to determine which line the cursor is
+  // on, we need this offset to correctly render and move the cursor.
+  Offset _floatingCursorOffset(TextPosition textPosition) =>
+      Offset(0, renderEditor.preferredLineHeight(textPosition) / 2);
+
   @override
   void updateFloatingCursor(RawFloatingCursorPoint point) {
-    throw UnimplementedError();
+    switch (point.state) {
+      case FloatingCursorDragState.Start:
+        if (floatingCursorResetController.isAnimating) {
+          floatingCursorResetController.stop();
+          onFloatingCursorResetTick();
+        }
+        // We want to send in points that are centered around a (0,0) origin, so
+        // we cache the position.
+        _pointOffsetOrigin = point.offset;
+
+        final currentTextPosition =
+            TextPosition(offset: renderEditor.selection.baseOffset);
+        _startCaretRect =
+            renderEditor.getLocalRectForCaret(currentTextPosition);
+
+        _lastBoundedOffset = _startCaretRect!.center -
+            _floatingCursorOffset(currentTextPosition);
+        _lastTextPosition = currentTextPosition;
+        renderEditor.setFloatingCursor(
+            point.state, _lastBoundedOffset!, _lastTextPosition!);
+        break;
+      case FloatingCursorDragState.Update:
+        assert(_lastTextPosition != null, 'Last text position was not set');
+        final floatingCursorOffset = _floatingCursorOffset(_lastTextPosition!);
+        final centeredPoint = point.offset! - _pointOffsetOrigin!;
+        final rawCursorOffset =
+            _startCaretRect!.center + centeredPoint - floatingCursorOffset;
+
+        final preferredLineHeight =
+            renderEditor.preferredLineHeight(_lastTextPosition!);
+        _lastBoundedOffset = renderEditor.calculateBoundedFloatingCursorOffset(
+          rawCursorOffset,
+          preferredLineHeight,
+        );
+        _lastTextPosition = renderEditor.getPositionForOffset(renderEditor
+            .localToGlobal(_lastBoundedOffset! + floatingCursorOffset));
+        renderEditor.setFloatingCursor(
+            point.state, _lastBoundedOffset!, _lastTextPosition!);
+        final newSelection = TextSelection.collapsed(
+            offset: _lastTextPosition!.offset,
+            affinity: _lastTextPosition!.affinity);
+        // Setting selection as floating cursor moves will have scroll view
+        // bring background cursor into view
+        renderEditor.onSelectionChanged(
+            newSelection, SelectionChangedCause.forcePress);
+        break;
+      case FloatingCursorDragState.End:
+        // We skip animation if no update has happened.
+        if (_lastTextPosition != null && _lastBoundedOffset != null) {
+          floatingCursorResetController
+            ..value = 0.0
+            ..animateTo(1,
+                duration: _floatingCursorResetTime, curve: Curves.decelerate);
+        }
+        break;
+    }
+  }
+
+  /// Specifies the floating cursor dimensions and position based
+  /// the animation controller value.
+  /// The floating cursor is resized
+  /// (see [RenderAbstractEditor.setFloatingCursor])
+  /// and repositioned (linear interpolation between position of floating cursor
+  /// and current position of background cursor)
+  void onFloatingCursorResetTick() {
+    final finalPosition =
+        renderEditor.getLocalRectForCaret(_lastTextPosition!).centerLeft -
+            _floatingCursorOffset(_lastTextPosition!);
+    if (floatingCursorResetController.isCompleted) {
+      renderEditor.setFloatingCursor(
+          FloatingCursorDragState.End, finalPosition, _lastTextPosition!);
+      _startCaretRect = null;
+      _lastTextPosition = null;
+      _pointOffsetOrigin = null;
+      _lastBoundedOffset = null;
+    } else {
+      final lerpValue = floatingCursorResetController.value;
+      final lerpX =
+          lerpDouble(_lastBoundedOffset!.dx, finalPosition.dx, lerpValue)!;
+      final lerpY =
+          lerpDouble(_lastBoundedOffset!.dy, finalPosition.dy, lerpValue)!;
+
+      renderEditor.setFloatingCursor(FloatingCursorDragState.Update,
+          Offset(lerpX, lerpY), _lastTextPosition!,
+          resetLerpValue: lerpValue);
+    }
   }
 
   @override
@@ -204,6 +294,20 @@ mixin RawEditorStateTextInputClientMixin on EditorState
     _textInputConnection!.connectionClosedReceived();
     _textInputConnection = null;
     _lastKnownRemoteTextEditingValue = null;
-    _sentRemoteValues.clear();
+  }
+
+  void _updateSizeAndTransform() {
+    if (hasConnection) {
+      // Asking for renderEditor.size here can cause errors if layout hasn't
+      // occurred yet. So we schedule a post frame callback instead.
+      SchedulerBinding.instance!.addPostFrameCallback((_) {
+        if (!mounted) {
+          return;
+        }
+        final size = renderEditor.size;
+        final transform = renderEditor.getTransformTo(null);
+        _textInputConnection?.setEditableSizeAndTransform(size, transform);
+      });
+    }
   }
 }
