@@ -1,3 +1,3 @@
 export 'bullet_point.dart';
-export 'checkbox.dart';
+export 'checkbox_point.dart';
 export 'number_point.dart';
