@@ -4,11 +4,81 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:image_picker/image_picker.dart';
 
-import '../../models/documents/nodes/embed.dart';
+import '../../models/documents/nodes/embeddable.dart';
+import '../../models/rules/insert.dart';
+import '../../models/themes/quill_dialog_theme.dart';
 import '../../translations/toolbar.i18n.dart';
-import '../../utils/media_pick_setting.dart';
+import '../../utils/platform.dart';
 import '../controller.dart';
 import '../toolbar.dart';
+
+class LinkDialog extends StatefulWidget {
+  const LinkDialog({this.dialogTheme, this.link, Key? key}) : super(key: key);
+
+  final QuillDialogTheme? dialogTheme;
+  final String? link;
+
+  @override
+  LinkDialogState createState() => LinkDialogState();
+}
+
+class LinkDialogState extends State<LinkDialog> {
+  late String _link;
+  late TextEditingController _controller;
+
+  @override
+  void initState() {
+    super.initState();
+    _link = widget.link ?? '';
+    _controller = TextEditingController(text: _link);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      backgroundColor: widget.dialogTheme?.dialogBackgroundColor,
+      content: TextField(
+        keyboardType: TextInputType.multiline,
+        maxLines: null,
+        style: widget.dialogTheme?.inputTextStyle,
+        decoration: InputDecoration(
+            labelText: 'Paste a link'.i18n,
+            labelStyle: widget.dialogTheme?.labelTextStyle,
+            floatingLabelStyle: widget.dialogTheme?.labelTextStyle),
+        autofocus: true,
+        onChanged: _linkChanged,
+        controller: _controller,
+      ),
+      actions: [
+        TextButton(
+          onPressed: _link.isNotEmpty &&
+                  AutoFormatMultipleLinksRule.linkRegExp.hasMatch(_link)
+              ? _applyLink
+              : null,
+          child: Text(
+            'Ok'.i18n,
+            style: widget.dialogTheme?.labelTextStyle,
+          ),
+        ),
+      ],
+    );
+  }
+
+  void _linkChanged(String value) {
+    setState(() {
+      _link = value;
+    });
+  }
+
+  void _applyLink() {
+    Navigator.pop(context, _link.trim());
+  }
+}
+
+enum MediaPickSetting {
+  Gallery,
+  Link,
+}
 
 class ImageVideoUtils {
   static Future<MediaPickSetting?> selectMediaPickSetting(
@@ -61,7 +131,7 @@ class ImageVideoUtils {
           'Please provide webImagePickImpl for Web '
           '(check out example directory for how to do it)');
       imageUrl = await webImagePickImpl!(onImagePickCallback);
-    } else if (_isMobile()) {
+    } else if (isMobile()) {
       imageUrl = await _pickImage(imageSource, onImagePickCallback);
     } else {
       assert(filePickImpl != null, 'Desktop must provide filePickImpl');
@@ -113,7 +183,7 @@ class ImageVideoUtils {
           'Please provide webVideoPickImpl for Web '
           '(check out example directory for how to do it)');
       videoUrl = await webVideoPickImpl!(onVideoPickCallback);
-    } else if (_isMobile()) {
+    } else if (isMobile()) {
       videoUrl = await _pickVideo(videoSource, onVideoPickCallback);
     } else {
       assert(filePickImpl != null, 'Desktop must provide filePickImpl');
@@ -125,8 +195,6 @@ class ImageVideoUtils {
       controller.replaceText(index, length, BlockEmbed.video(videoUrl), null);
     }
   }
-
-  static bool _isMobile() => Platform.isAndroid || Platform.isIOS;
 
   static Future<String?> _pickVideo(
       ImageSource source, OnVideoPickCallback onVideoPickCallback) async {
