@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
+import 'package:tuple/tuple.dart';
 
 import '../../models/documents/attribute.dart';
+import '../../models/rules/insert.dart';
 import '../../models/themes/quill_dialog_theme.dart';
 import '../../models/themes/quill_icon_theme.dart';
 import '../../translations/toolbar.i18n.dart';
 import '../controller.dart';
-import '../link_dialog.dart';
+import '../link.dart';
 import '../toolbar.dart';
-import 'quill_icon_button.dart';
 
 class LinkStyleButton extends StatefulWidget {
   const LinkStyleButton({
@@ -60,8 +61,8 @@ class _LinkStyleButtonState extends State<LinkStyleButton> {
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
-    final isEnabled = !widget.controller.selection.isCollapsed;
-    final pressedHandler = isEnabled ? () => _openLinkDialog(context) : null;
+    final isToggled = _getLinkAttributeValue() != null;
+    final pressedHandler = () => _openLinkDialog(context);
     return GestureDetector(
       onTap: () async {
         final dynamic tooltip = _toolTipKey.currentState;
@@ -83,7 +84,7 @@ class _LinkStyleButtonState extends State<LinkStyleButton> {
           icon: Icon(
             widget.icon ?? Icons.link,
             size: widget.iconSize,
-            color: isEnabled
+            color: isToggled
                 ? (widget.iconTheme?.iconUnselectedColor ??
                     theme.iconTheme.color)
                 : (widget.iconTheme?.disabledIconColor ?? theme.disabledColor),
@@ -97,18 +98,155 @@ class _LinkStyleButtonState extends State<LinkStyleButton> {
   }
 
   void _openLinkDialog(BuildContext context) {
-    showDialog<String>(
+    showDialog<dynamic>(
       context: context,
       builder: (ctx) {
-        return LinkDialog(dialogTheme: widget.dialogTheme);
+        final link = _getLinkAttributeValue();
+        final index = widget.controller.selection.start;
+
+        var text;
+        if (link != null) {
+          // text should be the link's corresponding text, not selection
+          final leaf =
+              widget.controller.document.querySegmentLeafNode(index).item2;
+          if (leaf != null) {
+            text = leaf.toPlainText();
+          }
+        }
+
+        text ??= widget.controller.document
+            .getPlainText(index, widget.controller.selection.end - index);
+        return _LinkDialog(
+            dialogTheme: widget.dialogTheme, link: link, text: text);
       },
     ).then(_linkSubmitted);
   }
 
-  void _linkSubmitted(String? value) {
-    if (value == null || value.isEmpty) {
-      return;
+  String? _getLinkAttributeValue() {
+    return widget.controller
+        .getSelectionStyle()
+        .attributes[Attribute.link.key]
+        ?.value;
+  }
+
+  void _linkSubmitted(dynamic value) {
+    // text.isNotEmpty && link.isNotEmpty
+    final String text = (value as Tuple2).item1;
+    final String link = value.item2.trim();
+
+    var index = widget.controller.selection.start;
+    var length = widget.controller.selection.end - index;
+    if (_getLinkAttributeValue() != null) {
+      // text should be the link's corresponding text, not selection
+      final leaf = widget.controller.document.querySegmentLeafNode(index).item2;
+      if (leaf != null) {
+        final range = getLinkRange(leaf);
+        index = range.start;
+        length = range.end - range.start;
+      }
     }
-    widget.controller.formatSelection(LinkAttribute(value));
+    widget.controller.replaceText(index, length, text, null);
+    widget.controller.formatText(index, text.length, LinkAttribute(link));
+  }
+}
+
+class _LinkDialog extends StatefulWidget {
+  const _LinkDialog({this.dialogTheme, this.link, this.text, Key? key})
+      : super(key: key);
+
+  final QuillDialogTheme? dialogTheme;
+  final String? link;
+  final String? text;
+
+  @override
+  _LinkDialogState createState() => _LinkDialogState();
+}
+
+class _LinkDialogState extends State<_LinkDialog> {
+  late String _link;
+  late String _text;
+  late TextEditingController _linkController;
+  late TextEditingController _textController;
+
+  @override
+  void initState() {
+    super.initState();
+    _link = widget.link ?? '';
+    _text = widget.text ?? '';
+    _linkController = TextEditingController(text: _link);
+    _textController = TextEditingController(text: _text);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      backgroundColor: widget.dialogTheme?.dialogBackgroundColor,
+      content: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          const SizedBox(height: 8),
+          TextField(
+            keyboardType: TextInputType.multiline,
+            style: widget.dialogTheme?.inputTextStyle,
+            decoration: InputDecoration(
+                labelText: 'Text'.i18n,
+                labelStyle: widget.dialogTheme?.labelTextStyle,
+                floatingLabelStyle: widget.dialogTheme?.labelTextStyle),
+            autofocus: true,
+            onChanged: _textChanged,
+            controller: _textController,
+          ),
+          const SizedBox(height: 16),
+          TextField(
+            keyboardType: TextInputType.multiline,
+            style: widget.dialogTheme?.inputTextStyle,
+            decoration: InputDecoration(
+                labelText: 'Link'.i18n,
+                labelStyle: widget.dialogTheme?.labelTextStyle,
+                floatingLabelStyle: widget.dialogTheme?.labelTextStyle),
+            autofocus: true,
+            onChanged: _linkChanged,
+            controller: _linkController,
+          ),
+        ],
+      ),
+      actions: [
+        TextButton(
+          onPressed: _canPress() ? _applyLink : null,
+          child: Text(
+            'Ok'.i18n,
+            style: widget.dialogTheme?.labelTextStyle,
+          ),
+        ),
+      ],
+    );
+  }
+
+  bool _canPress() {
+    if (_text.isEmpty || _link.isEmpty) {
+      return false;
+    }
+
+    if (!AutoFormatMultipleLinksRule.linkRegExp.hasMatch(_link)) {
+      return false;
+    }
+
+    return true;
+  }
+
+  void _linkChanged(String value) {
+    setState(() {
+      _link = value;
+    });
+  }
+
+  void _textChanged(String value) {
+    setState(() {
+      _text = value;
+    });
+  }
+
+  void _applyLink() {
+    Navigator.pop(context, Tuple2(_text.trim(), _link.trim()));
   }
 }
